# front-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OOUN Restaurant API Documentation",
        "description": "University Restaurant Management System API - OOUN Tunisia",
        "contact": {
            "name": "OOUN Restaurant System",
            "email": "admin@ooun.tn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/feedback": {
            "get": {
                "tags": [
                    "Student - Feedback"
                ],
                "summary": "Get student feedback history",
                "operationId": "20193cdb7ac1efc479f528e27cdec6ca",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Student - Feedback"
                ],
                "summary": "Submit feedback",
                "description": "Submit feedback for a menu and earn 10 points",
                "operationId": "60736e8377bf9cf8c7bf569139ad37f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "menu_id",
                                    "rating",
                                    "category"
                                ],
                                "properties": {
                                    "menu_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 4
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "taste",
                                            "quality",
                                            "service",
                                            "cleanliness"
                                        ],
                                        "example": "taste"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Great meal!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Feedback submitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/feedback/{id}": {
            "get": {
                "tags": [
                    "Student - Feedback"
                ],
                "summary": "Get specific feedback",
                "operationId": "0917424b7b70c55b6c9cc5bbfb39d53b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations": {
            "get": {
                "tags": [
                    "Student - Reservations"
                ],
                "summary": "Get all student reservations",
                "description": "Retrieve paginated list of authenticated student's reservations",
                "operationId": "46f3beef2bd4b5baab51116b6e45657f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Student - Reservations"
                ],
                "summary": "Create new reservation",
                "operationId": "5cf1a2b9f5c353f06dc56180f4c2e7a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "menu_id",
                                    "date",
                                    "payment_method"
                                ],
                                "properties": {
                                    "menu_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "wallet",
                                            "points",
                                            "cash"
                                        ],
                                        "example": "wallet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created"
                    },
                    "400": {
                        "description": "Insufficient balance or unavailable menu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{id}": {
            "get": {
                "tags": [
                    "Student - Reservations"
                ],
                "summary": "Get specific reservation",
                "operationId": "baab38a9b1cf5d546c3542f47c990884",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{id}/cancel": {
            "post": {
                "tags": [
                    "Student - Reservations"
                ],
                "summary": "Cancel reservation",
                "description": "Cancel a reservation and process refund if applicable",
                "operationId": "fc6df3ebd89dd21ded6f5053ca30df55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled successfully"
                    },
                    "400": {
                        "description": "Already cancelled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}